consolidator tests: token is minted twice with same ID
